feat: página Home criada. Rotas e redirecionamentos corrigidos para '/home'.
