+++ src/components/HomeSecretaria/index.js
@@ -1,0 +1,2 @@
+import HomeSecretaria from './HomeSecretaria';
+export default HomeSecretaria;
