--- src/pages/HomeSecretaria/index.js
@@ -1,2 +1,0 @@
-import HomeSecretaria from './HomeSecretaria';
-export default HomeSecretaria;
