+++ src/routes.js
@@ -1,28 +1,27 @@
-import React from "react";
-import { BrowserRouter, Route, Switch, Redirect } from "react-router-dom";
-import AlterarSenha from "./pages/AlterarSenha";
-import AlterarSenhaComEmail from "./pages/AlterarSenhaComEmail";
-import AlterarSenhaEsquecida from "./pages/AlterarSenhaEsquecida";
-import AreaReceitas from "./pages/AreaReceitas";
-import Cadastro from "./pages/Cadastro";
-import Chat from "./pages/Chat";
-import CriacaoFormulario from "./pages/CriacaoFormulario";
-import EdicaoConteudo from "./pages/EdicaoConteudo";
-import Header from "./components/Header/Header";
-import HomeMedico from "./pages/HomeMedico/HomeMedico";
-import HomeSecretaria from "./pages/HomeSecretaria/HomeSecretaria";
-import ListaFormularios from "./pages/ListaFormularios";
-import ListaUsuarios from "./pages/ListaUsuarios";
-import Login from "./pages/Login";
-import ModeloReceitas from "./pages/ModeloReceitas";
-import Perfil from "./pages/Perfil";
-import PerfilPaciente from "./pages/PerfilPaciente";
-import RespostaFormulario from "./pages/RespostaFormulario";
-import EditarPerfil from "./pages/EditarPerfil";
-import Agendamentos from "./pages/Agendamentos";
-import { usuarioAutenticado, recebeTipo } from "./services/auth";
-import EditarFormulario from "./pages/EditarFormulario/EditarFormulario";
-import FormularioEspecifico from "./pages/FormularioEspecifico";
+import React from 'react';
+import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
+import AlterarSenha from './pages/AlterarSenha';
+import AlterarSenhaComEmail from './pages/AlterarSenhaComEmail';
+import AlterarSenhaEsquecida from './pages/AlterarSenhaEsquecida';
+import AreaReceitas from './pages/AreaReceitas';
+import Cadastro from './pages/Cadastro';
+import Chat from './pages/Chat';
+import CriacaoFormulario from './pages/CriacaoFormulario';
+import EdicaoConteudo from './pages/EdicaoConteudo';
+import Header from './components/Header/Header';
+import ListaFormularios from './pages/ListaFormularios';
+import ListaUsuarios from './pages/ListaUsuarios';
+import Login from './pages/Login';
+import ModeloReceitas from './pages/ModeloReceitas';
+import Perfil from './pages/Perfil';
+import PerfilPaciente from './pages/PerfilPaciente';
+import RespostaFormulario from './pages/RespostaFormulario';
+import EditarPerfil from './pages/EditarPerfil';
+import Agendamentos from './pages/Agendamentos';
+import { usuarioAutenticado, recebeTipo } from './services/auth';
+import EditarFormulario from './pages/EditarFormulario/EditarFormulario';
+import FormularioEspecifico from './pages/FormularioEspecifico';
+import Home from './pages/Home/Home';
 
 const RotasPrivadas = ({ component: Component, ...rest }) => (
   <Route
@@ -32,7 +31,7 @@ const RotasPrivadas = ({ component: Component, ...rest }) => (
         <Component {...props} />
       ) : (
         <Redirect
-          to={{ pathname: "/login", state: { from: props.location } }}
+          to={{ pathname: '/login', state: { from: props.location } }}
         />
       )
     }
@@ -42,12 +41,12 @@ const RotasPrivadasMedico = ({ component: Component, ...rest }) => (
   <Route
     {...rest}
     render={(props) =>
-      usuarioAutenticado() && recebeTipo() === "MASTER" ? (
+      usuarioAutenticado() && recebeTipo() === 'MASTER' ? (
         <Component {...props} />
       ) : (
         <Redirect
           to={{
-            pathname: "/web/homemedico",
+            pathname: '/web/home',
             state: { from: props.location },
           }}
         />
@@ -60,11 +59,14 @@ const RotasPrivadasSecretaria = ({ component: Component, ...rest }) => (
   <Route
     {...rest}
     render={(props) =>
-      usuarioAutenticado() && recebeTipo() === "SECRETARIA(O)" ? (
+      usuarioAutenticado() && recebeTipo() === 'SECRETARIA(O)' ? (
         <Component {...props} />
       ) : (
         <Redirect
-          to={{ pathname: "/web/homesecretaria", state: { from: props.location } }}
+          to={{
+            pathname: '/web/home',
+            state: { from: props.location },
+          }}
         />
       )
     }
@@ -75,18 +77,26 @@ function Routes() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route exact path="/login" component={Login} />
-        <Route exact path="/fc43c2dd-cf6c-4807-825e-3c9e7ba41b19e19637d2-5a44-463c-bae5-6709e7e53448/alterarsenha" component={AlterarSenhaEsquecida}/>
-        <Route exact path="/alterarsenha_requisicao" component={AlterarSenhaComEmail} />
-        <RotasPrivadas exact path="/cadastro" component={Cadastro} />
+        <Route exact path='/login' component={Login} />
+        <Route
+          exact
+          path='/fc43c2dd-cf6c-4807-825e-3c9e7ba41b19e19637d2-5a44-463c-bae5-6709e7e53448/alterarsenha'
+          component={AlterarSenhaEsquecida}
+        />
+        <Route
+          exact
+          path='/alterarsenha_requisicao'
+          component={AlterarSenhaComEmail}
+        />
+        <RotasPrivadas exact path='/cadastro' component={Cadastro} />
         <RotasPrivadasMedico
           exact
-          path="/web"
-          component={() => <Redirect to="/web/listadeusuarios" />}
+          path='/web'
+          component={() => <Redirect to='/web/listadeusuarios' />}
         />
-        <RotasPrivadas path="/web" component={UserHeader} />
+        <RotasPrivadas path='/web' component={UserHeader} />
         <RotasPrivadasMedico
-          component={() => <Redirect to="/web/listadeusuarios" />}
+          component={() => <Redirect to='/web/listadeusuarios' />}
         />
       </Switch>
     </BrowserRouter>
@@ -99,81 +109,74 @@ function UserHeader() {
       <Switch>
         <RotasPrivadas
           exact
-          path="/web/editarperfil"
+          path='/web/editarperfil'
           component={EditarPerfil}
         />
         <RotasPrivadas
           exact
-          path="/web/alterarsenha"
+          path='/web/alterarsenha'
           component={AlterarSenha}
         />
         <RotasPrivadasMedico
-          path="/web/areareceitas"
+          path='/web/areareceitas'
           component={AreaReceitas}
         />
         <RotasPrivadasSecretaria
           exact
-          path="/web/agendamentos"
+          path='/web/agendamentos'
           component={Agendamentos}
         />
-        <RotasPrivadas exact path="/web/chat" component={Chat} />
+        <RotasPrivadas exact path='/web/chat' component={Chat} />
         <RotasPrivadasMedico
           exact
-          path="/web/criacaoformulario"
+          path='/web/criacaoformulario'
           component={CriacaoFormulario}
         />
         <RotasPrivadasMedico
           exact
-          path="/web/editarformulario"
+          path='/web/editarformulario'
           component={EditarFormulario}
         />
         <RotasPrivadasMedico
           exact
-          path="/web/editarconteudo"
+          path='/web/editarconteudo'
           component={EdicaoConteudo}
         />
         <RotasPrivadas
           exact
-          path="/web/listaformularios"
+          path='/web/listaformularios'
           component={ListaFormularios}
         />
         <RotasPrivadas
           exact
-          path="/web/listadeusuarios"
+          path='/web/listadeusuarios'
           component={ListaUsuarios}
         />
         <RotasPrivadasMedico
           exact
-          path="/web/modeloreceitas"
+          path='/web/modeloreceitas'
           component={ModeloReceitas}
         />
-        <RotasPrivadas exact path="/web/perfil" component={Perfil} />
+        <RotasPrivadas exact path='/web/perfil' component={Perfil} />
         <RotasPrivadas
           exact
-          path="/web/perfildopaciente"
+          path='/web/perfildopaciente'
           component={PerfilPaciente}
         />
         <RotasPrivadas
           exact
-          path="/web/respostaformulario"
+          path='/web/respostaformulario'
           component={RespostaFormulario}
         />
         <RotasPrivadas
           exact
-          path="/web/formularioespecifico"
+          path='/web/formularioespecifico'
           component={FormularioEspecifico}
         />
-        <RotasPrivadasSecretaria
-          exact
-          path="/web/homesecretaria"
-          component={HomeSecretaria}
+        <RotasPrivadas exact path='/web/home' component={Home} />
+        <RotasPrivadas
+          component={() => <Redirect to='/web/listadeusuarios' />}
         />
-        <RotasPrivadasMedico
-          exact
-          path="/web/homemedico"
-          component={HomeMedico}
-        />
-        <RotasPrivadas component={() => <Redirect to="/web/listadeusuarios" />} />
       </Switch>
     </Header>
   );
